Moved custom Exception subclasses to new file
