+++ syscore/exceptions.py
@@ -1,0 +1,15 @@
+"""
+Custom exceptions
+"""
+
+
+class missingContract(Exception):
+    pass
+
+
+class missingData(Exception):
+    pass
+
+
+class existingData(Exception):
+    pass
