+++ syscore/objects.py
@@ -217,16 +217,6 @@ def get_class_name(class_object):
     return class_object.__name__
 
 
-class missingData(Exception):
-    pass
-
-
-class missingContract(Exception):
-    pass
-
-class existingData(Exception):
-    pass
-
 missing_file = named_object("missing file")
 missing_data = named_object("missing data")
 ALL_ROLL_INSTRUMENTS = "ALL"
